cas: clean blockTransform, panic intentionally
There was a bug in diskv (https://github.com/peterbourgon/diskv/pull/20)
which meant that the blockTransform function could be passed an empty
string. In our case this just paniced after trying to split + access
parts of the resulting empty slice.

Panicking is arguably correct behaviour (if we ever get something that's
not a hash it's likely a bug in our code), but let's be more intentional
about it.