pkg/tar: add overwrite option to ExtractTar.
Add an overwrite option to ExtractTar so, calling the function on a non empty
directory, will overwrite existing files instead of returning an error.

It's needed during ACI rendering, as (in the actual implementation), extracting
a dependency list of ACIs over the same directory, the files provided by the
next ACI in the list have to overwrite the ones previously extracted.