stage1/init: quote Exec* argument lists
Note this only quotes the arguments, not the path.  There's a bug in
systemd which prevents quoting the path; it thinks the executable's path
is not absolute when quoted.  This prevents things like an executable path
containing spaces from working:

https://bugs.freedesktop.org/show_bug.cgi?id=86171

The fix landed in systemd upstream Dec 18th, which should make it into v219.

Also added keys to the unit.UnitOption composite literals to satisfy govet.

Fixes #321.