rkt: refactor fetch, add tests
Unfortunately the recent fetch refactoring inverted the logic from what
it should be; we want to first check if the supplied image name is a
valid URL, and, if so, we should just pass it through. Otherwise,
newDiscoveryApp will return bogus results because
discovery.NewAppFromString does not expect to get URLs.

Also adds a few test cases for NewDiscoveryApp.