cas: remote: clarify the meaning of Blob and rename it to BlobKey.
As Remote.Blob is used to store the related key on the diskv blob store, rename
it to BlobKey.

It should also be noted that it's not the same as an ImageID (as an ImageID can
be a sha512 with different key lenghts) and to deriver a
BlobKey from an imageID one should use cas.ResolveKey(ImageID).