rkt: introduce the volume flag
Introduce a more complete volume format that can specify arguments and details
of the volume to be mounted. For example:

```
rkt run --volume database,kind=host,source=/tmp
```

This will implicitly create a mount points for every app called 'database'. In
a follow-on patch we will introduce a --mount flag that will add a particular
volume to a particular app.