stage0: do not panic on missing app
Now that apps are an optional part of the ImageManifest, we need to
check an image contains one before accessing it.

At some point it would be nicer to check this higher up the stack, but
that will require reworking the interface between `rkt run` and stage0,
so for now just do it here.