Godeps: bump appc/spec, remove go-iptables
Bump appc to 0.3.0+git. Until we reach v1.0.0, we are going to keep
Rocket releases trailing spec releases so the versioning is not too
confusing.

The reason for removing go-iptables is that nothing currently references
it, so Godep does not think it should be vendored: see
https://github.com/coreos/rocket/pull/470

It can be re-added when a package within Rocket uses it again.