cas: add a db handler.
This db handler wraps a database/sql and handles (but it can be changed) a
github.com/cznic/ql db adding external locking functions and various utility
functions.

As multiple rkt processes can concurrently try to access the db there the need
to lock it. This is done locking the dir containing the db file.

To avoid long wait times of concurrent rkt processes trying to access a DB,
only short living transactions should be done.

The DoTx function executes the provided txfuncs inside a unique transaction.
If one of the functions returns an error the whole transaction is rolled back.

The Do functions calls DoTx but before it locks and opens the db, and after it
closes and unlocks the db.

For the reasons explained above the Do function should always be used.