cas: add new ImageManifest diskv store and save ImageManifest on WriteACI
Add a new diskv store to save the ImageManifest of an ACI. As future functions
will need the image manifest this avoids the need (and slowness) of extracting
it from the ACI every time.

As an ACI and its image manifest are tied together, on WriteACI the image
manifest is saved using the same key of the blob store.