cas: check key prefix and min size in ResolveKey.
Using a very short key like "sha512-a", cas.blockTransform will panic
retrieving parts[1][0:2] as part[1] will be "a" and [0:2] is out of range. As
blockTransform cannot return an error this should be checked before calling
diskv functions.

Assuming that all cas functions using diskv should be called passing a full
key (if needed retrieving it with cas.ResolveKey), this patch just adds
additional validations to cas.ResolveKey for the right prefix and a minimal
length.

A more paranoid patch should verify this before any call to diskv functions.