rkt: support overlay filesystem
This mounts stage1 and the application images as an overlay filesystem
using each ACI's cached tree as the lower filesystem.

Also, the mounts are done in a separate mount namespace so they will be
unmounted when the container exits and they're not visible by the rest
of the system.

Systems that don't support overlay fall back to plain copying.