cas: Introduce latest pattern handling for image fetching.
On fetching, if downloaded with the "latest" pattern (no version label
specified), the ACI is saved in the store with the "latest" flag.

In future, on "rkt fetch" or "rkt run", a way for the user to define if the
provided images (by local file or URL) should be saved in the store with the
lastest flag, should be provided. For doing this there's the need to find how
the user can request this for a specific file/URL as multiple files/URls can be
provided.