cas: add image locking.
This patch adds image and tree store locking like explained in #460.
An image is composed by the ACI file in the blob store and it's imageManifest
in the imageManifest store.
This means that a lock on an image covers multiple elements.
For this reason the keyLock functions are used to take a lock.

To avoid deadlocks the image locks shouldn't be taken inside the db access
code.

The tree store is kept separate from the image for various reasons:

* It can be composed by multiple images if the upper image has dependencies
* It's rendered and removed separately from image.

For this reason also the tree store locks are different from the image locks.
The tree store locks should be taken before the image locks.