rkt: fix enter when running without overlay fs
When not using overlay fs, the tree cache was not getting populated.
Since we now take the enter binary from the tree cache, this was
breaking rkt enter.

We now populate stage1's tree cache in every case. We don't do it for
the app images because it can add a significant amount of time in the
first run if the image is big and it's not needed unless you use overlay
fs.