rkt: add -local flag to run/prepare.
This patch adds a -local flag to use only local images (if available) from the
local store.

If the provided image name is a URL-like app name (like coreos.com/etcd:2.0.9)
then discovery is disabled and the image is searched in the store using
cas.GetACI.

If the provided image name has an "http", "https", "docker" scheme it will avoid
downloading if it's not in the store (not existing in the remote table).

This can be useful to use rkt in offline mode or to just simply avoid
downloading.