rkt: sync prepared pod file contents on disk.
before returning a prepared pod uuid ensure that its contents are synced to
disk. This avoid problems running an inconsistent pod if the machine crashes
after preparing and before running a pod.

This is done in two steps:

* Before calling p.xToPrepared sync all contents on disk. By now syncfs is
used.
* After calling p.xToPrepared sync the PreparedDir (the parent dir of all
prepared pods) to ensure that the directory renaming is saved in its parent dir
metadata.

Additionaly, the above parent directory sync, is done in every xTo* function.

`rkt run` probably doesn't needs this as the pods are executed just after
preparing and if the machine crashes the pod becomes garbage.

This implementation needs real world tests to verify that, under typical
workloads, no big slowdowns are introduced in the prepare phase due to the
syncfs call.