rkt: fetch images dependencies.
After fetching an image also fetch recursively all its dependencies.
Dependencies needs to be resolved using discovery as the unique useful
informations (as per spec) are the app name and labels.

By now, the finder in rkt/run.go and rkt/prepare.go and the fetcher in
rkt/fetch.go are configured with withDep = true for all images except the
stage1.

In a future, if there's a need, an option can be added to disable dependencies
fetching.