build: ability to choose a different systemd version
When building stage1 from sources with RKT_STAGE1_USR_FROM=src,
it is now possible to choose which systemd tag or branch should be
built:

  $ export RKT_STAGE1_USR_FROM=src
  $ export RKT_STAGE1_SYSTEMD_SRC=git://anongit.freedesktop.org/systemd/systemd
  $ export RKT_STAGE1_SYSTEMD_VER=v215
  $ ./build

By default, it still fetch the upstream git repository with tag v215 as
before. So defining the new variables is not necessary.

It is also possible to build against a local systemd repository:

  $ export RKT_STAGE1_USR_FROM=src
  $ export RKT_STAGE1_SYSTEMD_SRC=/home/user/git/systemd
  $ export RKT_STAGE1_SYSTEMD_VER=HEAD
  $ ./build

This should help developers to test newer systemd versions.

The patches in stage1/rootfs/usr_from_src/patches/ are versioned: the
subdirectory used is defined by $RKT_STAGE1_SYSTEMD_VER. If "HEAD" is
used, the master patchset will be selected.