test: Compute testable and/or formattable directories
Basically we want to run go test on all directories where _test.go
files reside excluding functional tests and Godeps. I'm rather not
sure about the last one - it might be useful to check if our Godeps
pass their tests.

We also want to check for formatting and suspicious constructs on all
.go files.

So instead of hardcoding the list, employ some basic tools to produce
the list for us. That way we don't have to maintain the list - nobody
ever remembers about that thing.