functional tests: use the correct version of go when using sudo
The functional tests use sudo because the tests need to be run as root.
Sudo does not preserve $PATH. When the user use a home-built go, it
might have a different version thatn the system-wide installed go.  It
causes problems when linking the tests with github.com/coreos/rkt/common

This patch fixes the issue by using the go environment of the user.