stage0: touch status directory when using overlay fs
When using overlayfs, we read the status directory from its upper layer
since the fs might be unmounted or in another mount namespace. If the
directory is not modified it won't be in the upper layer causing the
status command to fail.

We touch the directory to force the status directory to be in the upper
layer of the overlayfs.