rkt/rmimage: Apply reference check to the whole image.
By now only the treestore isn't removed if an image is referenced by
some pod.
This brings the need of a store gc to remove the rendered images in
the treestore when they aren't referenced anymore. This is difficult to
coordinate and is an extra work that can be avoided simply not removing
also the image from the store (store.RemoveACI) if the image is
referenced. IMHO it's also more consistent as the image doesn't
disappear from the `rkt images` output.