stage1: add halt.target
When systemd-nspawn receives a TERM signal it will send it to systemd in
stage1 which will try to run the halt.target. This target was not
present in the current stage1 so the container was not being shut down.

Fix this by adding halt.target which calls reaper.service to shut down
the container cleanly.