stage1: fix metadata service registration errors
When registering a pod to the metadata service, if individual app
registrations failed then an error object was being instantiated but not
actually returned. This fixes it to return the error and also to make a
best-effort attempt to unregister the pod in this scenario.

Also tweaks the file operations to close manifests more immediately
after they've been used.

Also fixes minor typo in rkt/metadata_service error message.