stage1: flavor src: print systemd version with git-describe
When a test fails because of systemd, it is useful to know which version
of systemd was used exactly. But this information is not always known:
when stage1 is build with systemd from git-master, we don't know exactly
which commit was used for the build because git-master moves often.

Write the git version both in the build logs (useful for continuous
integration) and in stage1.aci.

Unfortunately, git-describe does not always work well on repository
cloned with --depth so I am removing the --depth optimization.