rkt: Add custom help.
This patch changes the default cobra help/usage with a custom one that
resembles the original rkt usage output.

cobra has the concept of help and usage. The cobra help, by default writes the
short/long (if available) command description followed by the command
usage. The command usage contains the invocation line, the subcommands,
local and global flags and additional help topic (help for every command
not having an associated run function but having a subcommand as it will
not appear in the commands list).

This patch makes help and usage print the same output, additionally it
writes, for every command usage, also all the subcommands at every level
(not just the subcommands at the first level).