rkt/run: added --set-env-file switch and priorities for environments
--set-env-file gets an environment variables file path in the format "VAR=VALUE\n...".
This switch may be used multiple times on the command line to provide multiple
environment files. When it used multiple times, it merges each environment file
in a single environment and it returns an error if the same variable is set multiple
times (either in different files or in the same file).

This patch implements priorities among the various way to specify
environment variables, from highest to least priority:
1. Command line (--set-env)
2. From files (--set-env-file)
3. Image Manifest
4. Inherit (--inherit-env)
A variable in a group with an higher priority will override the value
of a variable with the same name in groups with lower priority without
producing any error. Variables with unique names will be always preserved.

Fixes https://github.com/coreos/rkt/issues/647