rkt/images.go: fix fetchImageFrom().
If the image is a file path (scheme == 'file'), then we bypass
the on-disk store.

If the '--local' flag is provided, then we will only fetch from
the on-disk store (unless image is a file path).

If the label is 'latest', then we will bypass the on-disk store
(unless '--local' is specified).

Otherwise if '--local' is false, image is not a file path, and the
lable is not 'latest' or empty, we will first try to fetch from the
on-disk store, if not found, then fetch from the internet.