rkt: fix --private-net without arguments
starting from spf13/pflag#20 it accepts also `--flag value` while
before `--flag=value` was needed. Additionally it supports default values
for flags specified without an explicit value (`--flag`) (which is
different from the default value for flags not passed as arguments).

In that case, the `--flag=value` is needed for flags that can have a
default value because the parser won't know if the argument after
`--flag` is the flag's value or another argument.

That fix also changed the way flags are handled and removed the use of
the boolFlag interface as now default values can be provided for every
type (not just bool) by setting the NoOptDefVal variable.

This patch fixes rkt to use the new behavior.