Move pod registration from stage1 to stage0
With token based auth, there's no longer a dependency
on networking (which is set up in stage1). It makes sense
to do registration in stage0 as it's a generic operation.
Also, registration is now performed by default. It can be
disabled by --mds-regiser=false. Unregistration is handled
by the GC.