fetch: check signature identity before downloading
This is an optimization to avoid having to download the image twice when
the user forgets to use "rkt trust".

Note that this optimization does not help when downloading the signature
returns the HTTP error 202 (such as quay.io on the first conversion).

Fixes: https://github.com/coreos/rkt/issues/573#issuecomment-78361925

Fixes partially, but not completely: https://github.com/coreos/rkt/issues/776