build: Do so-libs copying as an opt-in from flavor
Currently only "src" flavor opts in for copying so libs from
host. "host" flavor gets all required libraries at runtime. "coreos"
flavor is supposed to be self-contained, so it does not opt-in for
copying.

Also, fix the copying itself, as it should copy only libraries that
are not in ACI rootfs.