functional tests: Run GC on cleanup and reset
Now that the overlay fs is not mounted in a separate mount namespace
mounts will stay in host's mount namespace and to unmount them we need
to run the GC. Tests were failing because we tried to remove the
temp directory with a mounted filesystem inside.

This commit runs the GC before removing the temp directory.