stage1: add ctrl-alt-del and poweroff targets
These targets are called by machinectl reboot and poweroff respectively.
In ctrl-alt-del we just reboot the system without saving the exit
statuses of the apps and the poweroff target is just the same as the
halt.target: it calls the reaper to save the exit statuses of the apps
and shuts down the container.