build: Rewrite depsgen in go
This commit:
1. Removes dependency on Perl and Capture::Tiny module.
2. Simplifies the generated dep.mk file.
3. Makes dep.mk file generation slightly faster.

This commit also fixes a bug where a go project wasn't rebuilt when
its source files was modified. The bug existed because of my
misunderstanding of intermediate files. Intermediate files are ignored
if they have no prerequisites and the file that depend on it already
exists. Just use phony target to invalidate a project.

Using phony target uncovered another bug - using wildcard to get a
list of .go files inside a directory was also getting files which were
ignored by go during build (those with _test or _unsupported or _arm
suffixes and so on), so that output was almost always different from
the output of go list. Right now, generated dep.mk will call "go list"
for go projects instead of using wildcard.

For plain lists of files, wildcards are still used.