build: Add key-value depsgen
This mode basically takes an even number of arguments where an odd one
is a variable name and the following one is its value. During
subsequent make execution, when the generated deps file is included,
the current value of the variable is compared to the one stored in
generated deps file. If they are not equal, the target (passed via
--target parameter) is invalidated.