selinux: check if selinuxPath exists.
When rkt is compiled with selinux support (selinux tag) but selinux is disabled
the selinux mount point doesn't exists. Actually it's not checked so rkt thinks
selinux is enabled. It will then fail mounting overlayfs with selinux labels.

This patch checks for the existence of /sys/fs/selinux.

Note: This is different than the check done in
https://github.com/opencontainers/runc/blob/master/libcontainer/selinux/selinux.go#L55
that needs an additional package.