functional tests: add socket activation test
We add a small echo program that's compatible with systemd socket activation.

The test creates a pair of socket and service unit files for a container
running this program and starts the socket unit file which makes systemd
on the host listen on a port on behalf of the service.

Then we test socket activation by connecting to this port, sending a
string and checking that we get that same string. This will start the
container and pass the connection to the program inside seamlessly.