build: Add GOARCH_FOR_BUILD
When generating tools to use during the build they should be built for the build
machine, and not the target.

Add a section to configure.ac that sets up the make variable GOARCH_FOR_BUILD
which holds the value that GOARCH should be set to when building go language
build tools.  Also change depsgentool.mk to use GOARCH_FOR_BUILD.

Fixes errors like these when cross building:

  /bin/bash: tools/depsgen.tmp: No such file or directory

Signed-off-by: Geoff Levand <geoff@infradead.org>