rkt: fix fetchImageFromStore with missing os/arch labels.
Some images imported in the store will not contain os/arch labels.
This will happen with os/arch independent images.

In this case fetchImageFromStore will fail because it retrieves the image name and
labels from the image name string calling newDiscoveryApp that by default
adds the os and arch label for the current os and arch.

Adding the current os and arch labels is correct for the discovery phase and
should be used only there.

Instead `rkt image cat-manifest/export/extract/renders` works as their logic
doesn't add a default os and arch label.

This patch unifies and uses the same logic also for fetchImageFromStore.