store: print error reporting name and labels in GetACI.
Report provided name and labels in GetACI instead of a generic "aci not found"
error if it doesn't find an ACI matching them.

To get an ordered labels list, a new function labelsToString is added. It sorts
by label name giving precedence to the "version", "os" and "arch" (in this
order) labels.

Perhaps this function, if deemed useful, should be ported to appc/spec.