build: add default stage1 name and version
We add new configure options for the default name and version of the
stage1 aci. These are passed to rkt so it knows under which name and
version to search for stage1 in the store.

By default they are "coreos.com/rkt/stage1" and the rkt version number.
If packagers want to use a custom stage1 they should set these options
to the right values.