private-net test: fix error handling in goroutines
Use GoroutineAssistant to handle the following situation:
Goroutines created during a testrun should not use the Fatalf (Fail) functions
directly.  Instead they need to delegate the call to the goroutine that runs the
test. Further respect that created goroutines are not stopped immeditately by a
Fatalf in the main goroutine by returning the goroutine after an error is
detected.