build: Add make functions for undefining variables
They can be used to undefine all the variables in given namespaces in
a single line instead of creating long lists of empty value
assignments at the end of mk files or functions. This will be a bit of
a relieve, because there will be no need to keep those lists in sync
with newly-added or just-removed variables.