rkt: create a real temp dir for docker2aci images
We were writing the output file of docker2aci.Convert to the same temp
directory. Since the file name for a converted image is always the same
this was causing file corruption when fetching the same  docker image
with rkt several times in parallel.

We fixed that issue upstream in
https://github.com/appc/docker2aci/pull/76

In addition to the upstream fix, this patch makes the temporary
directory where rkt saves the converted docker image different in each
execution.