kvm: tun: give the network interface a name
Example of name: rkt-7543-tap0

The name contains "rkt" so it could be used by NetworkManager or ConnMan
to know it should ignore it.
See discussion in https://github.com/coreos/rkt/issues/1302

It contains 4 characters from the pod UUID. That should make it easier
to debug which network device is used by a pod. It does not contain the
full UUID because the device name is limited to 16 characters.