functional tests: Properly cleanup patched ACI in fetch tests
PatchTestACI must be followed up by the deferred image
removal. Otherwise we might clutter the disk with patched ACIs.

In this case the removal wasn't deferred, but it happened right after
importing it to CAS. The removal might not happen if the import
failed.

Replace the code with the newly introduced function that does the same
and makes sure that the patched ACI is indeed removed.