functional tests: Properly cleanup patched ACI in pod manifest tests
PatchTestACI must be followed up by the deferred image removal.
Otherwise we might clutter the disk with patched ACIs.

In this case the removal wasn't deferred, but it happened at the end
of the test case. The removal might not happen if the test failed.

Just use patchImportAndFetchHash function to get the desired result.