stage1: fix garbage collection after unmount
If we use overlayfs and we unmount the directories before
running rkt gc we won't have access to the stage1 directory but we need
the pod's flavor to do the correct GC (regular or LKVM).

We need to store the pod's flavor in the pod's root directory, trying
first in stage1 to keep backwards compatibility. This means it will
garbage collect old pods that weren't unmounted but if they did get
unmounted it will still fail.