build: Add a --glob-mode parameter to depsgen glob
Depsgen was by default generating a dep.mk files that was gather all
the files from a list of directories by using "*<suffix>" and
".*<suffix>" globs.

This behaviour might not be always suitable. For example for
generating dependencies on patches. Typical way of applying a set of
patches in make recipe was something like:

```
for p in patches_dir/*.patch; do
    patch $p ...
done
```

Even if in "patches_dir" there is a patch like ".hidden.patch" among
others, the "patches_dir/*.patch" glob will not pick it up. But the
generated dep.mk will surely do. So, removing ".hidden.patch" should
not affect building, but thanks (or not) to dep.mk it actually will
affect it.

Passing --glob-mode=normal to depsgen will generate a dep.mk without
the dependency on ".hidden.patch".